backend/requirements.txt
Flask==2.1.1
Flask-CORS==3.0.10
tensorflow
numpy==1.22.4
Pillow==9.1.1
tqdm==4.64.0
pandas==1.4.2
h5py==3.7.0
